% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dayLessons}
\alias{dayLessons}
\title{Exemplu de lecții, cu 64 profesori (între care, patru cuplaje dintre care două fac parte din tuplaje) și 34 clase}
\format{
\describe{
\item{prof}{Cod de 3 sau 6 caractere, reprezentând un profesor,
respectiv un cuplaj (doi profesori, pe grupe ale unei clase);
primele două litere abreviază disciplina pe care este încadrat
profesorul, iar cifra care urmează indexează profesorii
de pe aceeași disciplină}
\item{cls}{Clasa (două sau trei caractere) la care va intra
profesorul/cuplajul în ziua respectivă}
}
}
\usage{
dayLessons
}
\description{
set de 205 lecții prof|cls, care trebuie să se desfășoare
în orele 1:7 ale unei zile. Fiecare profesor are cel puțin una și
cel mult 7 ore (lecții); fiecare clasă are cel puțin 5 și cel mult 7 ore.
Unele lecții (indicate în setul "dayTuples") trebuie să cadă într-o aceeași
oră a zilei.
}
\keyword{datasets}
