% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dayTuples}
\alias{dayTuples}
\title{Exemplu de tuplaje}
\format{
\describe{
\item{prof}{Profesorii (eventual, cuplați doi pe o aceeași clasă) care
trebuie să intre într-o aceeași oră a zilei la câte una dintre clasele
respective (separați printr-un spațiu).}
\item{cls}{Clasele care trebuie tuplate (separate printr-un spațiu).}
}
}
\usage{
dayTuples
}
\description{
set de 4 tuplaje asociate setului "LSS". Un tuplaj
va angaja într-o aceeași oră, mai mulți profesori și mai multe clase, dar
un tuplaj corect trebuie să aibă același număr de profesori ca și de clase.
Dacă numărul de profesori este totuși mai mare decât al claselor din tuplaj,
utilizatorul va ști cel mai bine care dintre profesorii respectiv trebuie
cuplați, încât tuplajul să devină corect.
}
\keyword{datasets}
