% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_twins.R
\name{get_twins}
\alias{get_twins}
\title{Depistează cuplajele și dependențele de alocare pe ore, induse de acestea}
\usage{
get_twins(LSS)
}
\arguments{
\item{LSS}{data.frame cu lecțiile prof|cls}
}
\value{
NULL dacă nu există cuplaje; altfel, lista care indică bilateral,
pe fiecare membru al unui cuplaj și pe fiecare cuplaj, profesorii
și cuplajele de care depinde alocarea pe ore a lecțiilor sale.
}
\description{
Lecțiile cuplajelor trebuie să nu se suprapună (într-o aceeași oră) cu
lecțiile membrilor acestora.
}
\keyword{internal}
