% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LSS}
\alias{LSS}
\title{Exemplu de lecții, cu 62 profesori (între care, două cuplaje) și 32 clase}
\format{
\describe{
\item{prof}{Cod de 3 sau 6 caractere, reprezentând un profesor,
respectiv un cuplaj (doi profesori, pe grupe ale unei clase);
primele două litere abreviază disciplina pe care este încadrat
profesorul, iar cifra care urmează indexează profesorii
de pe aceeași disciplină}
\item{cls}{Clasa (două sau trei caractere) la care va intra
profesorul/cuplajul în ziua respectivă}
}
}
\usage{
LSS
}
\description{
set numit "LSS" de 204 lecții prof|cls, care trebuie să se desfășoare
într-o zi. În ziua respectivă, fiecare profesor are cel mult 7 lecții;
fiecare clasă are cel puțin 4 și cel mult 7 lecții.
}
\keyword{datasets}
