% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount_hours.R
\name{mount_hours}
\alias{mount_hours}
\title{Adaugă 'ora' încât oricare două lecții prof|cls|ora să nu se suprapună,
iar lecțiile dintr-un același tuplaj să cadă într-o aceeași oră 1:7}
\usage{
mount_hours(LSS, TPL = NULL)
}
\arguments{
\item{LSS}{data.frame cu lecțiile prof|cls, unde 'prof' este un
profesor propriu-zis, sau unul fictiv (cuplaj de doi/clasă)}

\item{TPL}{data.frame pentru tuplaje, dacă este cazul
Un tuplaj conține 2 sau mai mulți profesori, pe 2 sau mai multe clase
(numărul de profesori fiind egal cu al claselor)}
}
\value{
Un orar 'prof|cls|ora' pentru ziua respectivă
}
\description{
Adaugă 'ora' încât oricare două lecții prof|cls|ora să nu se suprapună,
iar lecțiile dintr-un același tuplaj să cadă într-o aceeași oră 1:7
}
\details{
'PERM' is a list of permutations for 4:7 elements.
'h2bin' is a vector of integers 2^(h-1) for the hours h=1:7.
'cnt_holes()' gives the current number of gaps in an individual schedule.
}
\examples{
\donttest{
    STP <- as.data.frame(mount_hours(dayLessons, dayTuples)) \%>\% 
           dplyr::arrange(cls, ora)
}
}
