% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tuplaje}
\alias{Tuplaje}
\title{Exemplu de tuplaje}
\format{
\describe{
\item{prof}{Profesorii care trebuie să intre într-o aceeași oră a zilei
la câte una dintre clasele respective (separați printr-un spațiu).}
\item{cls}{Clasele care trebuie tuplate (separate printr-un spațiu).}
}
}
\usage{
Tuplaje
}
\description{
set numit "Tuplaje", conținând 4 tuplaje asociate setului "LSS". Un tuplaj
va angaja într-o aceeași oră, mai mulți profesori și mai multe clase.
Lecțiile tuplate NU sunt înregistrate în "LSS".
}
\keyword{datasets}
