% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2matrix.R
\name{verify_matrix}
\alias{verify_matrix}
\title{Verifică matricea-orar}
\usage{
verify_matrix(MP)
}
\arguments{
\item{MP}{Matricea-orar}
}
\value{
TRUE dacă fiecare clasă apare cel mult o dată, pe fiecare coloană
(FALSE în caz contrar)
}
\description{
Fiecare clasă trebuie să apară cel mult o singură dată, pe fiecare coloană.
}
