% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hours2lessons-package.R
\docType{package}
\name{hours2lessons-package}
\alias{hours2lessons}
\alias{hours2lessons-package}
\title{hours2lessons: Alocă Pe Ore Lecțiile Zilei}
\description{
Lecțiile prof/cls trebuie completate cu un câmp "ora", astfel ca oricare două lecții prof/cls/ora să nu se suprapună într-o aceeași oră. The prof/cls lessons must be completed with a "hour" field ('ora), so that any two prof/cls/ora lessons do not overlap in the same hour. \url{https://vlad.bazon.net/}.
}
\author{
\strong{Maintainer}: Vlad Bazon \email{vlad.bazon@gmail.com}

}
\keyword{internal}
