% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2matrix.R
\name{long2matrix}
\alias{long2matrix}
\title{Transformă prof|cls|ora în matricea profesorilor pe orele 1:7}
\usage{
long2matrix(ORR)
}
\arguments{
\item{ORR}{data.frame prof|cls|ora (format lung)}
}
\value{
matricea orară a lecțiilor: clasele alocate pe ore fiecărui profesor
}
\description{
Transformă prof|cls|ora în matricea profesorilor pe orele 1:7
}
\examples{
long2matrix(as.data.frame(mount_hours(LSS, Tuplaje))) \%>\% print(quote=FALSE)

}
