% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long2matrix.R
\name{long2matrix}
\alias{long2matrix}
\title{Transformă prof|cls|ora în matricea profesorilor pe orele 1:7}
\usage{
long2matrix(ORR)
}
\arguments{
\item{ORR}{data.frame prof|cls|ora (format lung)}
}
\value{
matricea orară a lecțiilor: clasele alocate pe ore fiecărui profesor
}
\description{
"Matricea-orar" are liniile numite după profesori, fiecare linie conținând
clasele la care intră profesorul respectiv, în orele 1:7. Fiecare clasă
apare câte o singură dată pe fiecare coloană de rang mai mic sau egal cu
numărul de ore al clasei respective.
}
\examples{
\donttest{
    LM <- long2matrix(as.data.frame(mount_hours(LSS, Tuplaje)))
}
}
