% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mount_hours.R
\name{mount_hours}
\alias{mount_hours}
\title{Adaugă 'ora' încât oricare două lecții prof|cls|ora să nu se suprapună}
\usage{
mount_hours(LSS, TPL = NULL)
}
\arguments{
\item{LSS}{data.frame cu lecțiile prof|cls, unde 'prof' este un
profesor propriu-zis, sau unul fictiv (cuplaj de doi/clasă)}

\item{TPL}{data.frame pentru tuplaje, dacă este cazul
Un tuplaj conține 2 sau mai mulți profesori, pe 2 sau mai multe clase
(numărul de profesori fiind cel mult cu 1 mai mare, ca al claselor)}
}
\value{
Un orar prof|cls|ora pentru ziua respectivă
}
\description{
Adaugă 'ora' încât oricare două lecții prof|cls|ora să nu se suprapună
}
\examples{
    STP <- as.data.frame(mount_hours(LSS, Tuplaje)) \%>\% 
           dplyr::arrange(cls, ora)

}
