% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_prof_cls.R
\name{scale_prof_cls}
\alias{scale_prof_cls}
\title{Scalează profesorii (clasele) după numărul de clase (profesori) în comun}
\usage{
scale_prof_cls(LSS)
}
\arguments{
\item{LSS}{data.frame cu lecțiile prof|cls}
}
\value{
lista de coeficienți betweenness pentru profesori și clase
}
\description{
Clasele vor fi parcurse (pentru a monta ore 1:7 pe lecțiile fiecăreia)
în ordinea crescătoare a coeficienților "betweenness" din graful în care
două clase sunt adiacente după numărul de profesori comuni lor.
}
\keyword{internal}
