% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tuplaje}
\alias{Tuplaje}
\title{Exemplu de tuplaje}
\format{
set numit "Tuplaje", conținând 4 tuplaje asociate setului "LSS". Un tuplaj
va angaja într-o aceeași oră, mai mulți profesori și mai multe clase.
\describe{
\item{cls}{Clasele care trebuie tuplate}
\item{prof}{Profesorii care trebuie să intre într-o aceeași oră a zilei
la câte una dintre clasele respective. Dacă sunt mai mulți profesori
decât clase, se vor constitui cuplaje pe câte o clasă.}
}
}
\usage{
Tuplaje
}
\description{
Exemplu de tuplaje
}
\keyword{datasets}
